chore: Extract test-case management to doctest/parts/test_case.h
